pyproject.toml
[project]
name = "rragent"
version = "0.1.0"
description = "RRCLAW — unified AI agent harness for A-share quantitative trading"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "pagliazi"},
]
keywords = [
    "ai-agents",
    "multi-agent",
    "hermes-agent",
    "openclaw",
    "redis",
    "harness",
    "tool-calling",
    "agentic-ai",
    "quantitative-trading",
    "a-share",
    "self-learning",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "redis>=5.0",
    "websockets>=12.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "anthropic>=0.39",
    "aiohttp>=3.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]
dashscope = [
    "openai>=1.0",
]
full = [
    "openai>=1.0",
    "aiohttp>=3.9",
]

[project.scripts]
rragent = "rragent.__main__:main"
rrclaw-mcp = "rragent.tools.mcp.server:main"
rrclaw-market = "rragent.tools.mcp.reachrich_server:main"

[project.urls]
Homepage = "https://github.com/pagliazi/rrclaw"
Repository = "https://github.com/pagliazi/rrclaw"
Issues = "https://github.com/pagliazi/rrclaw/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["rrclaw"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
